fix : 색 태그 (Color)와 분위기 태그 (Mood)를  -> 무드 태그(Mood) 안의 enum(COLOR / VIBE)로 변경
