+++ backend/src/main/java/ku/hackerthon/BeMyMood/domain/mood/Mood.java
@@ -1,13 +1,17 @@
-package ku.hackerthon.BeMyMood.domain.tag;
+package ku.hackerthon.BeMyMood.domain.mood;
 
 import javax.persistence.*;
 
 @Entity
 public class Mood {
+
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "mood_id")
     private Long id;
 
     @Column(name = "mood_name")
     private String name;
+
+    @Enumerated(EnumType.STRING)
+    private MoodType type;
 }
