+++ backend/src/main/java/ku/hackerthon/BeMyMood/domain/mood/MoodType.java
@@ -1,0 +1,5 @@
+package ku.hackerthon.BeMyMood.domain.mood;
+
+public enum MoodType {
+    VIBE, COLOR
+}
