+++ backend/src/main/java/ku/hackerthon/BeMyMood/domain/spot/mood/SpotMood.java
